Fix: 품목명 select로 변경 && 입력 폼 형식 변경
입력 등의 작업을 위해서는 품목명을 알아야할 필요가 있음
input text에서는 품목명을 정확히 입력해야하므로
작업이 힘들기 때문에 select로 변경
입력 폼 형식 변경

+++ src/main/java/com/mit/model/EpleeDTO.java
@@ -8,6 +8,6 @@ import lombok.Setter;
 public class EpleeDTO extends UserDTO {
 	
 	private String email;
-	private Integer deptNum;
+	private Long deptNum;
 	
 }
